refactored particle system
