+++ junebugEngine/sprite.py
@@ -1,7 +1,6 @@
 import pygame
 import json
 import os.path
-import random
 
 
 # TODO: consider defining this elsewhere
@@ -132,9 +131,6 @@ class AnimSprite(pygame.sprite.Sprite):
             self.orientation = Orientation.RIGHT
         self.setAnimation(data['meta']['frameTags'][0]['name'])
 
-        # initialize list for particles
-        self.particles = []
-
     def die(self):
         """Plays the "die" animation, if available, then removes the sprite."""
         if not self.setAnimation('die'):
@@ -162,60 +158,6 @@ class AnimSprite(pygame.sprite.Sprite):
         Triggered whenever an animation finished playing."""
         if self.currentAnimation.name == "die":
             self.kill()
-
-    def emit_particles(self,
-                       color_list=[(255, 255, 255)],
-                       decay=0.1,
-                       x_velocity_range=[-1, 1],
-                       y_velocity_range=[-2, 2],
-                       radius_range=[1, 4]):
-        items_to_remove_ = []
-
-        # configuration (to be moved to function args)
-        decay = 0.1
-        x_scaled_range = x_velocity_range * 10
-        x_velocity = random.randint(x_scaled_range[0],
-                                    x_scaled_range[1]) / 10
-        y_velocity = random.randint(y_velocity_range[0],
-                                    y_velocity_range[1])
-        particle_radius = random.randint(radius_range[0],
-                                         radius_range[1])
-
-        # compute intervals for colors
-        interval_size = float(particle_radius) / float(len(color_list))
-        interval_borders = []
-        for ind in range(len(color_list)):
-            interval_borders.append(float(ind) * interval_size)
-        # create new particle
-        self.particles.append([list(self.rect.center),
-                               [x_velocity, y_velocity],
-                               particle_radius,
-                               particle_radius,
-                               color_list,
-                               interval_borders])
-        # compute existing particles
-        for particle in self.particles:
-            particle[0][0] += particle[1][0]
-            particle[0][1] += particle[1][1]
-            particle[2] -= decay
-            if particle[2] <= 0:
-                items_to_remove_.append(particle)
-        # remove nonexistent particles
-        for item in items_to_remove_:
-            self.particles.remove(item)
-
-    def draw_particles(self, offset, surface):
-        for particle in self.particles:
-            center = list(map(sum, zip(tuple(particle[0]), offset)))
-            # color stuff
-            for ind in range(len(particle[4])):
-                if particle[2] >= particle[5][ind]:
-                    color_index = ind
-
-            pygame.draw.circle(surface,
-                               particle[4][color_index],
-                               center,
-                               particle[2])
 
     def update(self, ms):
         """Updates the sprite by advancing the animation by [ms] ms."""
